% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlepackeR.R
\name{circlepackeR}
\alias{circlepackeR}
\title{htmlwidget for d3.js zoomable circle packing visualizations}
\usage{
circlepackeR(data, size = "size", color_min = "hsl(152,80\%,80\%)",
  color_max = "hsl(228,30\%,40\%)", width = NULL, height = NULL,
  color_col = NULL, tooltip = TRUE, tooltip_cols = NULL,
  quartile_values = NULL, quartile_colors = NULL, id_col = "id_col")
}
\arguments{
\item{data}{data in the form of a hierarchical list or a nested d3 JSON hierarchy}

\item{size}{string representing the name of the size variable.  \code{"size"} is
the default.}

\item{color_min}{string representing the minimum value of the color range for the
circles. The string can be either a hexadecimal, RGB, or HSL color.
\code{"hsl(152,80\%,80\%)"} is the default.}

\item{color_max}{string representing the maximum value of the color range for the
circles. The string can be either a hexadecimal, RGB, or HSL color.
\code{"hsl(228,30\%,40\%)"} is the default.}

\item{color_col}{string representing the name of the color column variable.  \code{"NULL"} is the default}

\item{tooltip}{boolean on whether to display tooltip for each node.  \code{"TRUE"} is the default.}

\item{tooltip_cols}{list of node property names to display in the tooltip. The default is to display all properties.}

\item{quartile_values}{list of the four quartile values to use as separate colors.}

\item{quartile_colors}{list of the four colors to use for the quartile values.}

\item{id_col}{string representing the name fo the unique identifier column variable \code{"id_col"} is the default}
}
\description{
\href{http://bl.ocks.org/mbostock/7607535}{Circle packing visualizations} provide
an interactive way of exploring hierarchical data that contains elements of different
sizes or magnitudes
}
\examples{
library(circlepackeR)

# using json
circlepackeR( "http://bl.ocks.org/mbostock/raw/7607535/flare.json" )

# using data.tree
\dontrun{
  library(data.tree)
  data(acme)
  circlepackeR(acme, size = "cost" )

}

# using a data frame with data.tree
\dontrun{
  library(data.tree)
  library(treemap)
  data(GNI2010)
  GNI2010$pathString <- paste("world",
                              GNI2010$continent,
                              GNI2010$country,
                              sep = "/")

  population <- as.Node(GNI2010)
  circlepackeR(population, size = "population" )

}
}
