% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circlepackeR.R
\name{circlepackeROutput}
\alias{circlepackeROutput}
\title{Widget output function for use in Shiny}
\usage{
circlepackeROutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

