% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circlepackeR.R
\name{renderCirclepackeR}
\alias{renderCirclepackeR}
\title{Widget render function for use in Shiny}
\usage{
renderCirclepackeR(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

