% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circlepackeR.R
\name{circlepackeR}
\alias{circlepackeR}
\title{htmlwidget for d3.js zoomable circle packing visualizations}
\usage{
circlepackeR(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{data in the form of a hierarchical list or a nested d3 JSON hierarchy}
}
\description{
\href{http://bl.ocks.org/mbostock/7607535}{Circle packing visualizations} provide
an interactive way of exploring hierarchical data that contains elements of different
sizes or magnitudes
}

