% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circlepackeR.R
\name{circlepackeR}
\alias{circlepackeR}
\title{htmlwidget for d3.js zoomable circle packing visualizations}
\usage{
circlepackeR(data, size = "size", color_min = "hsl(152,80\%,80\%)",
  color_max = "hsl(228,30\%,40\%)", width = NULL, height = NULL)
}
\arguments{
\item{data}{data in the form of a hierarchical list or a nested d3 JSON hierarchy}

\item{size}{string representing the name of the size variable.  \code{"size"} is
the default.}

\item{color_min}{string representing the minimum value of the color range for the
circles. The string can be either a hexadecimal, RGB, or HSL color.
\code{"hsl(152,80\%,80\%)"} is the default.}

\item{color_max}{string representing the maximum value of the color range for the
         circles. The string can be either a hexadecimal, RGB, or HSL color.
         \code{"hsl(228,30\%,40\%)"} is the default.}
}
\description{
\href{http://bl.ocks.org/mbostock/7607535}{Circle packing visualizations} provide
an interactive way of exploring hierarchical data that contains elements of different
sizes or magnitudes
}
\examples{
library(circlepackeR)

# using json
circlepackeR( "http://bl.ocks.org/mbostock/raw/7607535/flare.json" )

# using data.tree
\dontrun{
  library(data.tree)
  data(acme)
  circlepackeR(acme, size = "cost" )

}

# using a data frame with data.tree
\dontrun{
  library(data.tree)
  library(treemap)
  data(GNI2010)
  GNI2010$pathString <- paste("world",
                              GNI2010$continent,
                              GNI2010$country,
                              sep = "/")

  population <- as.Node(GNI2010)
  circlepackeR(population, size = "population" )

}
}

